Begin splitting out rsvp pages
