+++ src/App.css
@@ -228,25 +228,23 @@ nav ul {
   padding: 0;
   display: flex;
   list-style: none;
+  gap: 2em;
 }
 
 
 nav a {
-  margin: 0 1em;
-  margin-bottom: 2px;
   padding: 0.5em 0.25em;
+  color: var(--color-text);
+}
+
+nav li {
+  margin: 0 0 2px 0;
+  padding: 0;
   font-family: 'Noto Serif';
   font-style: italic;
   font-weight: 400;
   font-size: 20px;
   line-height: 27px;
-  transition: all 100ms ease-out;
-  color: var(--color-text);
-}
-
-nav li {
-  margin: 0;
-  padding: 0;
   transition: all 100ms ease-out;
 }
