+++ src/components/GuestList.css
@@ -10,6 +10,10 @@
     width: stretch;
 }
 
+#rsvpNav {
+    background-color: white;
+}
+
 #guest-input-wrapper {
     max-width: 400px;
     margin: 2em auto;
